Add ability to set/increase timeout for openoffice conversion. (#5853)
Include the whole hon_trie_lru_pqc test in the ifdef(PROPER) block so that macros are available in the whole file

Add documentation for getting kazoo to build on Mac OS

Fix link in mac documentation

Make the dependency installation clearer in the mac documentation

Add more dependencies to the mac installation documentation

Cleanup formatting. Add details to the mac os documentation to make running `make fmt` work.

Add the installing-on-mac md file to the documentation. Add some details to the doc about running `make docs`

Fix the header level on the installing mac docs

formatting and naming updates